requirements.txt
xecd-rates-client==1.0.0
apache-airflow==2.6.0
psycopg2-binary==2.9.9